Refactored the code responsible for extracting build structure from Bazel queries.
